changed CustomFunctions.php to camel case
new Collector rule: all php files that are not meant to be run by the end user are camel cased
